build/python/build/project.py: add "lazy" parameter to make_build_path()

+++ build/python/build/project.py
@@ -56,8 +56,10 @@ class Project:
             push_all(toolchain, path, self.patches)
         return path
 
-    def make_build_path(self, toolchain):
+    def make_build_path(self, toolchain, lazy=False):
         path = os.path.join(toolchain.build_path, self.base)
+        if lazy and os.path.isdir(path):
+            return path
         try:
             shutil.rmtree(path)
         except FileNotFoundError:
